Pin landuse column names to a fixed set
The user can no longer specify custom names in their csv files.
Instead, they must use the names that we provide in the documentation.
This makes it easier for us to specify the resampling techniques.
And we can pin to which wflow variables the landuse maps are written.

Rename of the landuse maps was moved to a protected function that is
reused by wflow_sbm.py

Fixes #430