Pin landuse column names to a fixed set (#623)
* Pin landuse column names to a fixed set

The user can no longer specify custom names in their csv files. Instead, they must use the names that we provide in the documentation. This makes it easier for us to specify the resampling techniques. And we can pin to which wflow variables the landuse maps are written.

Since the user can no longer have custom names, we can simply use these names to find the correct resampling method.

Fixes #430